feat(be): bigint/money config types + reserved-row enforcement + currency support
- Rename integer to bigint in config type vocabulary (DB patch + fire-and-forget)

- Add money config type with currency metadata in type_config

- Enforce reserved-row rules in BE wrapper (type/type_config locked if reserved)

- Add config-entries router tests

- Update environment-configuration user guide doc

- Update AGENTS.md with config table standard documentation

+++ db-meta/fire-and-forget/add_config_table_standard_columns.sql
@@ -169,7 +169,7 @@ UPDATE "public"."auth_configurations" SET
 WHERE "key" = 'auth_roles_path';
 
 UPDATE "public"."auth_configurations" SET
-  "type" = 'integer',
+  "type" = 'bigint',
   "type_config" = NULL,
   "label_key" = 'config.auth.invitation_expiry_days.label',
   "description_key" = 'config.auth.invitation_expiry_days.description'
@@ -204,7 +204,7 @@ UPDATE "public"."auth_configurations" SET
 WHERE "key" = 'redis_url';
 
 UPDATE "public"."auth_configurations" SET
-  "type" = 'integer',
+  "type" = 'bigint',
   "type_config" = NULL,
   "label_key" = 'config.auth.mfa_challenge_token_ttl_seconds.label',
   "description_key" = 'config.auth.mfa_challenge_token_ttl_seconds.description'
